backend/requirements.txt
flask
flask-cors
pandas
vaderSentiment
textblob
praw
apscheduler
python-dotenv
